requirements/dev.txt
-r base.txt

django-debug-toolbar==1.4
django-extensions==1.6.1
django-template-debug==0.3.5
